Add support for HSQLDB (use SampleData by default)
Since the Kettle endpoints have no way to know what kind of database they're
connected to, the ddlFile config option has been added. It points to the file
with the proper SQL DDL script for the database specified in 'jndi'.

The database schema is now different and incompatible with the previous one,
as it no longer uses PostgreSQL arrays but instead it uses an extra
'sections_roles' table. Arrays are largely non-standard, so using them and
supporting multiple DBs was not a good idea.